Rework StreamBuffer and make it the core transaction primitive
This reworks StreamBuffer, and makes it the core primitive
for transacting via read and write. It is now aware of the Device MTU,
and will align to the MTU. It also reworks the time stamps to align to
each transaction, supporting multiple time stamps per buffer for latency
analysis. We also allow the full vector type to be a Parameter,
to support other array types.