Ensure that `read!(s::Stream, ...)` does not loop forever
When the underlying hardware is not properly configured, we should not
loop forever in `read!()`, especially as we are given a `timeout` value.
If we time out, we should return.  We do not yet have a good way to
surface these kinds of errors to the user, we may need to provide a
different API (that returns error values, etc...) for more advanced
usage.