Check device/stream are open before `{read,write}Stream`
It's best to ensure that a given device or stream is open before
attempting to read/write to it.  This is especially important in the
event that you are doing something like:

```
Device(...) do dev
   @async read_loop(dev)
   ...
end
```

In the event that an exception is triggered, the `Device` will be
cleaned up by the `do`-block `Device` method, and the asynchronous
`read_loop()` will very quickly attempt to read from the now-defunct
`Device` object.  This segfaults, often bringing the process down before
the original exception has had a chance to print out.

This PR changes the behavior to instead pay a miniscule cost each time
we attempt to use the `Device` and `Stream` objects, translating
segfaults to exceptions.