Add `GC.@preserve` to `buffers` when calling `read!()`/`write!()`
Without this, there is a chance that the compiler will drop `buffers`
before SoapySDR has had a chance to read it in for transmission/write
into it for reception.  Note that this is _not_ only needed for
`write!()`, it is also necessary for `read!()` on the off-chance that we
have inlined `read!()` and thrown away the resulting buffer; the
optimizer is then free to do whatever it likes with `buffer`, including
free it before SoapySDR has finished with it.